Book and loan management

+++ src/main/java/com/rtrnonato/library_management/repositories/BookRepository.java
@@ -1,9 +1,7 @@
 package com.rtrnonato.library_management.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.rtrnonato.library_management.entities.Book;
 
-public interface BookRepository extends JpaRepository<Book, Integer>{
-
+public interface BookRepository extends JpaRepository<Book, Long>{
 }
